Move QCQMC code to recirq/qcqmc/conftest.py
Fixes #473.
